04_assignments/assignments_01/Project 3 QR code encoder  decoder Python Project/pyproject.toml
[project]
name = "project-3-qr-code-encoder-decoder-python-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "image>=1.5.33",
    "pillow>=11.1.0",
    "qrcode>=8.0",
]
